Scrub down unit tests, add basic checksumming
This moves all the file manipulation stuff into a single place in a
base class for our unit tests.

It also adds checksums in the process, along with testing that they
work.  To ensure this is working as expected, I added some unicode
text in a new file, which has different string and binary
representations.