LocalFile: add test case for asymmetric checksum knowledge
Which is a fancy way of saying: if a LocalFile doesn't have a
checksum, it can't tell if its checksum is better than another
LocalFile with one.  So, if we don't have a checksum, skip checking
them (unless forced).  However, if we have one, treat `None` on the
other LocalFile as a checksum miss.