Add dev environment setup to enable faster iteration on CLI
While getting started on the command-line interface for baycat, it
became clear that things would go a lot easier if the workflow was
closer to the expected one.  To get there, we did the following:

* Fixed up requirement handling in setup.py
* Deleted requirements.txt as it was now redundant
* Removed tests/__init__.py because it was unneeded and confused
  setuptools
* Added a development environment setup shell script
* Moved the makefile's init routine over to call that script
* Added a trivial dummy CLI just to get things going.
* Updated README.md with a section on how to get the environment up
  and running, along with what the Makefile supports.

The net result of all this is a much cleaner repo ready for the work
that comes next.