Minor unit test cleanup (removes logging!)
This makes two small tweaks to the unit test setup.  The first is that
we no longer need to do any wacky path mangling stuff with the new dev
environment, which is great.

The second small tweak is disabling all logging within unit tests via
the setUp() method.  There are currently no provisions for overriding
this disable, but we could easily add an environment variable which
allows them to be re-enabled if it's set.  I imagine that I'll be
adding that code within a few weeks, but I'll wait until I need it to
do so.