Major sync bugfix and refactor
This commit fixes a critical bug where dry run would request files to
be deleted.

But, more importantly, it is a huge refactor of the sync strategies.
We move almost all the logic up into the abstract base class, and
reduce the subclassess to implementing the actual mechanism only.
That is, the `_rm_file` method now just deletes a file; all logic
about dry runs etc is handed up in the base class.

A huge benefit of this change is that the sync implementation classes
are now a lot simpler and easier to reason about.  It also greatly
reduces the amount of surface area which needs to be tested in each
sync class.

This commit also adds operation counters, tracking how many times
those actual operations are called during a sync process.  This can be
used in unit tests to audit the sync class's execution of the computed
diff states.  This is going to be important as we add S3 support and
start doing things which can cost money.  Having test coverage greatly
reduces the likelihood that we accidentally commit code that performs
a complete sync every single run (chewing up bandwidth and racking up
usage fees at S3).