Make LocalFile.delta consider time resolution in diffs
This is needed because not all systems have nanosecond time
resolution, so two timestamps may be equal if they're not within a
nanosecond of each other.

This still feels a bit undercooked as a methodology for declaring two
timestamps as identical, but it's better than what we had for
generalization.