Get S3 round-trip working cleanly
This commit gets a first test case of taking local file out to S3,
then syncing them back down to a new directory locally.

Basically, this is the first working proof of concept of baycat's core
systems.  Which is exciting, but it's clear that there's more
refactoring work to be done on it.

I think this is a reasonably clean version of the state flows for sync
to succeed.  Now that it's working, I'm willing to step back and
review the current flow, and make sure I'm not missing something
stupid.