First pass at making PathSelector JSONSerDes
This commit allows for loading a manifest off disk and then updating it
without needing to re-do the path selection process.

This makes the PathSelector a JSONSerDes object, so you can keep the
manifest generation arguments available in the stored object.  It also
updates the Manifest to make use of this selector, and fixes up a few
cases in the unit tests where we were deserialzing the Manifests in a
way that probably shouldn't have worked.