(WIP) A first pass at writing sync on the CLI
This is the first pass at writing sync on the CLI. This is still very
much Work In Progress, but it's a big step, as all the little tweaks
of the last few days are finally coming together.

LocalToLocal transfers work as expected, and single uploads to the
mock S3 interface are working okay as well.

Summary of the changes by class:

* Manifest: counters for bytes uploaded/deleted in manifests (used for
  rsync-y speedup computation)

* SyncStrategy: verbose and dry_run flags are added in

* SyncStrategy: small thinko around marking mkdirs as transferred fix

* CLI: adds CLIImpl class with some basic unit testing