Move from .baycat_manifest to .baycat/manifest
This is the first step in a big shift, moving our metadata from a file
to a subdirectory with the metadata in a file there.  This enables us
to hang more kinds of metadata in that directory, such as a local
cache of a remote S3Manifest.  We currently re-download the S3Manifest
contents every time we sync up to S3, which is wasteful (and adds up
with 100MB manifests).

That said, this is is a small step towards the final goal.  More will
follow to take advantage of this, including moving S3 metadata into a
similar structure in that repo layout.