Updates to S3Manifest, and knock-on effects from that
This fixes a bug in S3Manifest where it would overwrite a file's
LocalFile-derived entry in the manifest with the (metadata-deficient)
S3-derived version every time update() was called on the S3Manifest
instance.

This also fixes a bug in Manifest where .baycat/ the directory would
be added to the manifest when it shouldn't be.

The S3Manifest filename constant was renamed over to be a "key"
instead of a "filename," which is a helpful reminder that filenames in
S3 are just keys in a giant dictionary-as-a-service.

LocalFile::delta() doesn't consider uid, gid, or mode if they are set to
None in self.  If we don't know the value, it can't be different.

When syncing from S3 to local, if we don't have a UID or GID, we just
use the current uid/gid.  This seems like reasonable behavior, though
it should only happen if you're pulling down an S3 directory that
isn't currently managed by baycat.

We are starting to prefer to use expand_path instead of using a
"*File" object's .path member variable.  It's better to only have them
rely on their relative path, instead of repeating the whole path every
time.  In the near future, we will be removing path from the
serialization, but for now, we just rewrote some error messages to
make them easier to read.

S3File now removes leading slashes from rel_path when creating
entries, and ensures that its mtime_ns value is an int, so it can be
used in the local filesystem if needed.